Push GitHub webhooks to PubSub
* Create /api/github/webhook-subscription
* Migrate most GitHub logic to the subscription
* Create GithubWebhookMessage for passing data to the subscription
* Cleanup tests
  * Use defined RepositorySlug from Config
  * Don't pass default values
