+++ app_dart/lib/src/model/proto/protos.dart
@@ -3,5 +3,6 @@
 // found in the LICENSE file.
 
 export 'internal/build_status_response.pb.dart';
+export 'internal/github_webhook.pb.dart';
 export 'internal/key.pb.dart';
 export 'internal/scheduler.pb.dart';
